Disabled Jump to Enhance Experience
I removed jumping functionality from the smallest form of the character because it felt too good to play that form.

There was a danger that the player didn't wanted to every return to a "fatter" form, so I removed jumping.

The objective of this is to motivate the player to stay in the middle form (for most of the time), as it will give the best control and precesion.